java chinese iter_2

+++ src/main/java/org/6367670c1a6d9265ec017a2a/6367670c1a6d9265ec017a2a.java
@@ -1,9 +1,7 @@
 public class ArrayConverter {
     /**
-     * <p>将对象类型的 Integer 数组转换为基本类型。</p> 
-     * <p>如果输入数组为 <code>null</code>，则此方法返回 <code>null</code>。</p>
-     * 
-     * @param array 一个 <code>Integer</code> 数组，可以为 <code>null</code>
+     * <p>将对象类型的 Integer 数组转换为基本类型。</p> <p>如果输入数组为 <code>null</code>，则此方法返回 <code>null</code>。</p>
+     * @param array  一个 <code>Integer</code> 数组，可以为 <code>null</code>
      * @return 一个 <code>int</code> 数组，如果输入数组为空则返回 <code>null</code>
      * @throws NullPointerException 如果数组内容为 <code>null</code>
      */
@@ -14,7 +12,7 @@ public class ArrayConverter {
         
         final int[] result = new int[array.length];
         for (int i = 0; i < array.length; i++) {
-            result[i] = array[i]; // Auto-unboxing, will throw NPE if array[i] is null
+            result[i] = array[i].intValue();
         }
         return result;
     }
